+++ src/main/java/cheng/community/controller/QuestionCotroller.java
@@ -1,8 +1,6 @@
 package cheng.community.controller;
 
 import cheng.community.dto.QuestionDTO;
-import cheng.community.mapper.QuestionMapper;
-import cheng.community.model.Question;
 import cheng.community.service.QuestionService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
@@ -19,7 +17,7 @@ public class QuestionCotroller {
     @Autowired
     private QuestionService questionService;
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name="id")Integer id
+    public String question(@PathVariable(name="id") Long id
                             , Model model){
         QuestionDTO questionDTO =questionService.getById(id);
         model.addAttribute("question",questionDTO);
